// generated/prisma/enums.ts
export const QuestionType = {
  SINGLE_CHOICE: 'SINGLE_CHOICE',
  TRUE_FALSE: 'TRUE_FALSE',
  SHORT_ANSWER: 'SHORT_ANSWER',
  ORDERING: 'ORDERING',
  NUMERIC: 'NUMERIC',
  PIN_ANCHOR: 'PIN_ANCHOR',
  SURVEY: 'SURVEY',
  WORD_CLOUD: 'WORD_CLOUD',
  DROP_PIN: 'DROP_PIN',
  LIKERT: 'LIKERT',
  SLIDE: 'SLIDE'
} as const

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType]